Directions for Hanoi.
I analyzed solutions online and found this. We can use it and it should work. Now the only thing left is to make a modification that WRITELN each of the steps that were used for this to happen. We must also write these instructions in PASCAL0 Syntax. 
I did some of the work. Now FIBB and making sure the compiler actually does what is says it does =)